+++ src/exchange/dmi.cpp
@@ -62,7 +62,7 @@ namespace internal{
       //std::ofstream ofile("dmi.txt");
 
       //	Loop over all atoms i
-      for(unsigned int i=0; i < atoms::num_atoms; i++){
+      for(unsigned int i=0; i < static_cast<unsigned int>(atoms::num_atoms); i++){
 
          // get material id for atom
          const unsigned int imat = atoms::type_array[i];
